fix clang build warnings and change return type

+++ src/internal.c
@@ -7636,7 +7636,7 @@ int CopyDecodedToX509(WOLFSSL_X509* x509, DecodedCert* dCert)
         while (cur != NULL) {
             if (cur->type == ASN_RFC822_TYPE) {
                 DNS_entry* dnsEntry;
-                int strLen = XSTRLEN(cur->name);
+                int strLen = (int)XSTRLEN(cur->name);
 
                 dnsEntry = (DNS_entry*)XMALLOC(sizeof(DNS_entry), x509->heap,
                                         DYNAMIC_TYPE_ALTNAME);
